fixes timeout logic and timeout page

+++ config/initializers/timeout.rb
@@ -1,2 +1,2 @@
-Rack::Timeout.timeout = Cfg.rack_execution_timeout.to_i
+Rack::Timeout.timeout = Cfg.rack_execution_timeout.to_f
